plotting code from Ione.
Here code that makes a 3d graph (and the matlab fig file). The x-axis
is the degree of counter-phaseness in the carrier gabors. The y axis
is the spacing of the carriers, and the z axis is the global
displacement needed to null the local motion (relative to the local
motion direction).

Each color is a different amount of counter-phaseness in the carrier
gabors

Each symbol is a different subject (with a couple of repeated symbols
that need fixing)

The blue line represents a pse of 0, which would imply that the local
motion had no effect on the direction of motion of the percept. Points
below the line imply that some global motion was needed in the
direction opposite to the local motion to null the percept. Points
above the line imply this ‘observer motion’ Peter has been noting.

What you see is that for spacings above 5 there is no effect of local
motion on the percept, below 5 you start seeing the effects of local
motion, which seems to lead to one of two percepts – either the
percept is biased in the direction of the local motion, or in the
induced opposite direction.

git-svn-id: svn+ssh://pat.shadlen.org/home/peterm/svn/eyetracking/analysis/ConcentricDirectionQuest@891 2a468891-300a-0410-9f4f-8ca473c5017f